requirements.txt
numpy
opencv-python
Pillow
llama-cpp-python
